+++ orchagent/aclorch.cpp
@@ -1415,115 +1415,6 @@ bool AclRuleMclag::validate()
     return true;
 }
 
-AclRulePbh::AclRulePbh(AclOrch *pAclOrch, string rule, string table, bool createCounter) :
-    AclRule(pAclOrch, rule, table, createCounter)
-{
-}
-
-bool AclRulePbh::validateAddPriority(const sai_uint32_t &value)
-{
-    SWSS_LOG_ENTER();
-
-    if ((value < m_minPriority) || (value > m_maxPriority))
-    {
-        SWSS_LOG_ERROR("Failed to validate priority: invalid value %d", value);
-        return false;
-    }
-
-    m_priority = value;
-
-    return true;
-}
-
-bool AclRulePbh::validateAddMatch(const sai_attribute_t &attr)
-{
-    SWSS_LOG_ENTER();
-
-    bool validate = false;
-
-    auto attrId = static_cast<sai_acl_entry_attr_t>(attr.id);
-    auto attrName = sai_serialize_enum(attrId, &sai_metadata_enum_sai_acl_entry_attr_t);
-
-    switch (attrId)
-    {
-        case SAI_ACL_ENTRY_ATTR_FIELD_GRE_KEY:
-        case SAI_ACL_ENTRY_ATTR_FIELD_IP_PROTOCOL:
-        case SAI_ACL_ENTRY_ATTR_FIELD_IPV6_NEXT_HEADER:
-        case SAI_ACL_ENTRY_ATTR_FIELD_L4_DST_PORT:
-        case SAI_ACL_ENTRY_ATTR_FIELD_INNER_ETHER_TYPE:
-            validate = true;
-            break;
-
-        default:
-            break;
-    }
-
-    if (!validate)
-    {
-        SWSS_LOG_ERROR("Failed to validate match field: invalid attribute %s", attrName.c_str());
-        return false;
-    }
-
-    m_matches[attrId] = attr.value;
-
-    return true;
-}
-
-bool AclRulePbh::validateAddAction(const sai_attribute_t &attr)
-{
-    SWSS_LOG_ENTER();
-
-    bool validate = false;
-
-    auto attrId = static_cast<sai_acl_entry_attr_t>(attr.id);
-    auto attrName = sai_serialize_enum(attrId, &sai_metadata_enum_sai_acl_entry_attr_t);
-
-    switch (attrId)
-    {
-        case SAI_ACL_ENTRY_ATTR_ACTION_SET_ECMP_HASH_ID:
-        case SAI_ACL_ENTRY_ATTR_ACTION_SET_LAG_HASH_ID:
-            validate = true;
-            break;
-
-        default:
-            break;
-    }
-
-    if (!validate)
-    {
-        SWSS_LOG_ERROR("Failed to validate action field: invalid attribute %s", attrName.c_str());
-        return false;
-    }
-
-    if (!AclRule::isActionSupported(attrId))
-    {
-        SWSS_LOG_ERROR("Action %s is not supported by ASIC", attrName.c_str());
-        return false;
-    }
-
-    m_actions[attrId] = attr.value;
-
-    return true;
-}
-
-bool AclRulePbh::validate()
-{
-    SWSS_LOG_ENTER();
-
-    if (m_matches.size() == 0 || m_actions.size() != 1)
-    {
-        SWSS_LOG_ERROR("Failed to validate rule: invalid parameters");
-        return false;
-    }
-
-    return true;
-}
-
-void AclRulePbh::update(SubjectType, void *)
-{
-    // Do nothing
-}
-
 bool AclTable::validateAddType(const acl_table_type_t &value)
 {
     SWSS_LOG_ENTER();
@@ -1632,6 +1523,10 @@ bool AclTable::create()
         table_attrs.push_back(attr);
 
         attr.id = SAI_ACL_TABLE_ATTR_FIELD_GRE_KEY;
+        attr.value.booldata = true;
+        table_attrs.push_back(attr);
+
+        attr.id = SAI_ACL_TABLE_ATTR_FIELD_ETHER_TYPE;
         attr.value.booldata = true;
         table_attrs.push_back(attr);
 
@@ -3423,13 +3318,13 @@ bool AclOrch::isCombinedMirrorV6Table()
 
 bool AclOrch::isAclMirrorTableSupported(acl_table_type_t type) const
 {
-    const auto &it = m_mirrorTableCapabilities.find(type);
-    if (it == m_mirrorTableCapabilities.cend())
+    const auto &cit = m_mirrorTableCapabilities.find(type);
+    if (cit == m_mirrorTableCapabilities.cend())
     {
         return false;
     }
 
-    return it->second;
+    return cit->second;
 }
 
 bool AclOrch::isAclActionSupported(acl_stage_type_t stage, sai_acl_action_type_t action) const
