+++ orchagent/pbh/pbhrule.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "aclorch.h"
+
+class AclRulePbh: public AclRule
+{
+public:
+    AclRulePbh(AclOrch *pAclOrch, string rule, string table, bool createCounter = false);
+
+    bool validateAddPriority(const sai_uint32_t &value);
+    bool validateAddMatch(const sai_attribute_t &attr);
+    bool validateAddAction(const sai_attribute_t &attr);
+    bool validate() override;
+    void update(SubjectType, void *) override;
+};
